+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Narrow-phase collision detection plugin (#524).
 - Ephemeral trait, used to indicate that relations and components shouldn't be persisted.
 - Collider gizmos plugin, used to view collision shapes.
+- Delta Time multiplier, can be used to adjust simulation speed (#866).
 
 ### Changed